# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kantin Rumah Kayu API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/api/v1/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "operationId": "00d3856d2d3d93124e0cc021d0cb5b79",
                "responses": {
                    "200": {
                        "description": "List user orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "operationId": "79da0a1adcc617490783dfd61927ff56",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tenant_id",
                                    "items",
                                    "type"
                                ],
                                "properties": {
                                    "tenant_id": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "pickup"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "menu_id",
                                                "qty"
                                            ],
                                            "properties": {
                                                "menu_id": {
                                                    "type": "integer"
                                                },
                                                "qty": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/api/v1/orders/{order}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "operationId": "38713af501dbd66af5a66bd568a49bc2",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/api/v1/orders/{order}/pay": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "operationId": "c9587aea02653c815193d092741625f9",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "paid_amount": {
                                        "type": "integer",
                                        "example": 24000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Marked paid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Already paid or invalid"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/api/v1/orders/{order}/pickup": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "operationId": "289f142d4188b9db1758661e26155a33",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marked picked up",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Not ready for pickup"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/api/v1/orders/{order}/complete": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "operationId": "bfdaa4b8d76bb40fcb24efddceb56c05",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marked completed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Not yet picked up"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Tenant": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Warung Nusantara"
                    },
                    "opens_at": {
                        "type": "string",
                        "example": "08:00:00"
                    },
                    "closes_at": {
                        "type": "string",
                        "example": "21:00:00"
                    }
                },
                "type": "object"
            },
            "Menu": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "tenant_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "example": "Nasi Goreng"
                    },
                    "price": {
                        "type": "integer",
                        "example": 25000
                    },
                    "category": {
                        "type": "string",
                        "example": "main"
                    },
                    "photo_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "stock": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
            "OrderItem": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "menu_id": {
                        "type": "integer"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "unit_price": {
                        "type": "integer",
                        "example": 12000
                    },
                    "subtotal": {
                        "type": "integer",
                        "example": 24000
                    },
                    "menu": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Menu"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Order": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "tenant_id": {
                        "type": "integer"
                    },
                    "tenant": {
                        "$ref": "#/components/schemas/Tenant"
                    },
                    "type": {
                        "type": "string",
                        "example": "pickup"
                    },
                    "status": {
                        "type": "string",
                        "example": "ready_for_pickup"
                    },
                    "payment_status": {
                        "type": "string",
                        "example": "paid"
                    },
                    "total_price": {
                        "type": "integer",
                        "example": 24000
                    },
                    "paid_amount": {
                        "type": "integer",
                        "example": 24000
                    },
                    "picked_up_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "completed_by_user": {
                        "type": "boolean",
                        "example": false
                    },
                    "completed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerToken": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}